pyproject.toml
[tool.poetry]
name = "qrcodegen"
version = "0.1.0"
description = ""
authors = ["lc3267 <melvin.strobl@kit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
qrcode = "^7.4.2"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
